♻ add usePost hook handling edit & create
